pyproject.toml
[tool.poetry]
name = "warehouse_management"
version = "0.1.0"
description = ""
authors = ["genagurbanguliyev <genagurbanguliyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
dependency-injector = "^4.42.0"
passlib = "^1.7.4"
psycopg = "^3.2.2"
python-jose = "^3.3.0"
rich = "^13.8.1"
asgi-correlation-id = "^4.3.3"
requests = "^2.32.3"
aiohttp = "^3.10.7"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-anyio = "^0.0.0"
httpx = "^0.27.2"
pytest-trio = "^0.8.0"
pytest-order = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
